pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "indyleg"
version = "0.1.0"
description = "Indiana Legal RAG Platform — AI-powered legal research and document intelligence"
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",

    # AWS
    "boto3>=1.34.0",
    "botocore>=1.34.0",

    # Config
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",

    # Document parsing
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",

    # Text / NLP
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    "tiktoken>=0.7.0",

    # Vector store (pgvector)
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.3.0",

    # OpenSearch (optional alternative)
    "opensearch-py>=2.6.0",

    # Async / queues
    "aioboto3>=13.0.0",
    "aio-pika>=9.4.0",

    # Auth
    "PyJWT>=2.8.0",

    # Observability
    "structlog>=24.2.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "moto[s3,sqs,batch]>=5.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S"]
ignore = [
    "S101",  # allow assert in tests
    "B008",  # allow Depends() in FastAPI defaults
    "S608",  # SQL from internal vars, not user input
    "UP017", # keep timezone.utc (local Python < 3.11)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S105", "S106"]  # test fixtures use dummy credentials

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
